More compiler warning fixes
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@34664 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/ptclib/httpclnt.cxx
@@ -424,7 +424,7 @@ bool PHTTPClient::ReadResponse(PMIMEInfo & replyMIME)
       PString body;
       if (lastResponseCode >= 300) {
 #if PTRACING
-        if (replyMIME.GetVar(ContentLengthTag(), numeric_limits<int64_t>::max()) <= MaxTraceContentSize)
+        if (replyMIME.GetVar(ContentLengthTag(), numeric_limits<int64_t>::max()) <= (int64_t)MaxTraceContentSize)
           ReadContentBody(replyMIME, body);
         else
 #endif
